create character sheet component
